6 Kyu Find the Parity Outlier solved. Took me a while. First to figure out how to check the parity. Use a pen and paper to get your algo down, or the whiteboard.
It passed the test cases but not submit, so I then had to learn how to figure out the problem. A crawl through issues taught how you should print to see the test case you failed. That' the commented lines

Found out that abs was needed because the remainder for -1 is given as -1 not 1. Sigh. Found that through the comments.

Of course the top solutions are nutty as usual. Makes me look like a chump.

fun find(integers: Array<Int>) = integers.singleOrNull { it % 2 == 0 } ?: integers.single { it % 2 != 0 }

and,

fun find(integers: Array<Int>): Int {
  val (even, odd) = integers.partition { it % 2 == 0 }
  return if (even.size == 1) even[0] else odd[0]
}

and,

fun find(integers: Array<Int>): Int =
    if( integers.take(3).count {it % 2 == 0} > 1 )
        integers.first {it % 2 != 0}
    else
        integers.first {it % 2 == 0}

 No brain to work through these rn, later